LayoutTests/inspector/model/parse-script-syntax-tree-expected.txt

passed AssignmentExpression
passed ArrayExpression
passed BlockStatement
passed BinaryExpression
passed BreakStatement
passed CallExpression
passed TryStatement
passed CatchClause
passed ConditionalExpression
passed ContinueStatement
passed DoWhileStatement
passed DebuggerStatement
passed EmptyStatement
passed ExpressionStatement
passed ForStatement
passed ForInStatement
passed ForOfStatement
passed FunctionDeclaration
passed FunctionExpression
passed Identifier
passed IfStatement
passed Literal
passed LabeledStatement
passed LogicalExpression
passed MemberExpression
passed NewExpression
passed ObjectExpression
passed Property
passed RestStatement
passed ReturnStatement
passed SequenceExpression
passed SwitchStatement
passed SwitchCase
passed ThisExpression
passed ThrowStatement
passed UnaryExpression
passed UpdateExpression
passed VariableDeclaration
passed VariableDeclarator
passed WhileStatement
passed WithStatement
passed YieldExpression
passed ClassStatement, Super, MetaProperty
passed AssignmentPattern
passed ArrowFunctionExpression
passed computed method on object literal
passed method on object literal
passed computed method property on object literal
passed ALL TESTS
